+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# Changelog for Unicode Set 0.13.0
+
+This is the changelog for Unicode Set 0.13.0 released on April 4th, 2021.  For older changelogs please consult the release tag on [GitHub](https://github.com/elixir-unicode/unicode_set/tags)
+
+## Enhancements
+
+* Adds `Unicode.Set.to_generated_match/1` that returns a tuple whose first element is the AST of a guard clause and the second element is a list of strings. This function is marked private and is implemented to suport [unicode_transform](https://hex.pm/unicode_transform) which uses this information to generate optimised code for matching unicode sets in a `case` expression.
+
 # Changelog for Unicode Set 0.12.0
 
 This is the changelog for Unicode Set 0.12.0 released on February 23rd, 2021.  For older changelogs please consult the release tag on [GitHub](https://github.com/elixir-unicode/unicode_set/tags)